% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelkappa.R
\name{modelkappa}
\alias{modelkappa}
\title{Function to calculate model-based kappa of agreement and association and their standard errors.}
\usage{
modelkappa(data, category, item, rater)
}
\arguments{
\item{data}{an optional data frame containing the variables providing \code{category}, \code{item}, and \code{rater}.}

\item{category}{a vector that identifies the ratings or results.}

\item{item}{a vector that identifies the items to be rated.}

\item{rater}{a vector that identifies the raters.}
}
\value{
Outputs the estimated summary measures containing
Number of observations, Number of categories, Number of items, Number of raters, Model-based kappa for agreement, its standard errors and 95% confidence intervals.
If number of categories is >2, then will also output Model-based kappa for association, its standard errors and 95% confidence intervals.
}
\description{
The \code{data} must be provided in case level or equivalently in `long' format that includes item identifiers, rater identifiers, and ratings (or categories).
}
\examples{
data(holmdata)
modelkappa(data=holmdata, cat=Cat, item=Item, rater=Rater)
}
\author{
Aya Mitani
}
